后端一顿乱写

+++ src/main/java/org/example/cyberforum/service/CommentService.java
@@ -1,27 +1,31 @@
 package org.example.cyberforum.service;
 
 import lombok.extern.slf4j.Slf4j;
+import org.example.cyberforum.dto.CommentInfo;
 import org.example.cyberforum.entities.Comment;
 import org.example.cyberforum.mapper.CommentMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import stark.dataworks.boot.autoconfig.web.LogArgumentsAndResponse;
+import stark.dataworks.boot.web.ServiceResponse;
 
 import java.util.Date;
 import java.util.List;
 
 @Slf4j
 @Service
+@LogArgumentsAndResponse
 public class CommentService
 {
     @Autowired
-    CommentMapper commentMapper;
+    private CommentMapper commentMapper;
 
     @Autowired
-    UserService userService;
+    private UserService userService;
 
     @Autowired
-    BlogService blogService;
+    private BlogService blogService;
 
 
 
@@ -32,44 +36,54 @@ public class CommentService
             return false;
 
         comment.setCreateTime(new Date());
-        log.info("CommentService addComment: add comment: " + comment);
+        log.info("CommentService addComment: add comment: {}", comment);
         commentMapper.addComment(comment);
         return true;
     }
 
-    public List<Comment> getCommentsByBlogId(Long blogId)
+    public ServiceResponse<List<CommentInfo>> getCommentsByBlogId(Long blogId)
     {
-        List<Comment> comments = commentMapper.getCommentsByBlogId(blogId);
-        for (Comment comment: comments)
-        {
-            comment.setUsername(userService.getUserById(comment.getUserId()).getUserName());
-        }
-        return comments;
+        if (!blogService.ifContainsBlogOfId(blogId))
+            return ServiceResponse.buildErrorResponse(-100, "blog not found");
+        List<CommentInfo> comments = commentMapper.getCommentInfoListByBlogId(blogId);
+        return ServiceResponse.buildSuccessResponse(comments);
     }
 
     // 分页
-    public List<Comment> getCommentsByBlogIdWithTop(Long blogId)
+    public ServiceResponse<List<CommentInfo>> getCommentsByBlogIdWithTop(Long blogId)
     {
-        List<Comment> comments = getCommentsByBlogId(blogId);
+        if (!blogService.ifContainsBlogOfId(blogId))
+            return ServiceResponse.buildErrorResponse(-100, "blog not found");
+
+        List<CommentInfo> comments = commentMapper.getCommentInfoListByBlogId(blogId);
         comments.sort((comment1, comment2) -> comment2.isTop() ? 1 : -1);
-        return comments;
+        return ServiceResponse.buildSuccessResponse(comments);
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void deleteCommentById(Long commentId)
+    public ServiceResponse<Boolean> deleteCommentById(Long commentId)
     {
+        if (!commentMapper.ifContainsComment(commentId))
+            return ServiceResponse.buildErrorResponse(-100, "comment not found");
         commentMapper.deleteCommentById(commentId);
+        return ServiceResponse.buildSuccessResponse(true);
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void putTop(Long commentId)
+    public ServiceResponse<Boolean> putTop(Long commentId)
     {
+        if (!commentMapper.ifContainsComment(commentId))
+            return ServiceResponse.buildErrorResponse(-100, "comment not found");
         commentMapper.putTop(commentId);
+        return ServiceResponse.buildSuccessResponse(true);
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void cancelTop(Long commentId)
+    public ServiceResponse<Boolean> cancelTop(Long commentId)
     {
+        if (!commentMapper.ifContainsComment(commentId))
+            return ServiceResponse.buildErrorResponse(-100, "comment not found");
         commentMapper.cancelTop(commentId);
+        return ServiceResponse.buildSuccessResponse(true);
     }
 }
